+++ upkie/envs/__init__.py
@@ -9,17 +9,13 @@ import logging
 import gymnasium as gym
 
 from .init_randomization import InitRandomization
-from .reward import Reward
-from .survival_reward import SurvivalReward
 from .upkie_base_env import UpkieBaseEnv
 from .upkie_wheeled_pendulum import UpkieWheeledPendulum
 
 __all__ = [
     "InitRandomization",
-    "Reward",
     "UpkieBaseEnv",
     "UpkieWheeledPendulum",
-    "SurvivalReward",
     "register",
 ]